feat(EarlySecurityCheckLLD): extract language change prompt into dedicated hook
